factor out test code
