+++ rest-api/test/unit_test/offline_test/public_metrics_export_test.py
@@ -20,9 +20,6 @@ from unit_test_util import FlaskTestBase, CloudStorageSqlTestBase, SqlTestBase, 
 from unit_test_util import make_questionnaire_response_json, PITT_HPO_ID
 
 TIME = datetime.datetime(2016, 1, 1)
-TIME_2 = datetime.datetime(2016, 1, 2)
-TIME_3 = datetime.datetime(2016, 1, 3)
-TIME_4 = datetime.datetime(2016, 1, 4)
 
 
 class PublicMetricsExportTest(CloudStorageSqlTestBase, FlaskTestBase):
@@ -37,38 +34,6 @@ class PublicMetricsExportTest(CloudStorageSqlTestBase, FlaskTestBase):
   def tearDown(self):
     super(PublicMetricsExportTest, self).tearDown()
     FlaskTestBase.doTearDown(self)
-
-  def submit_questionnaire_response(self, participant_id, questionnaire_id,
-                                    race_code, gender_code, state,
-                                    date_of_birth):
-    code_answers = []
-    date_answers = []
-    if race_code:
-      code_answers.append(('race', Concept(PPI_SYSTEM, race_code)))
-    if gender_code:
-      code_answers.append(('genderIdentity', Concept(PPI_SYSTEM, gender_code)))
-    if date_of_birth:
-      date_answers.append(('dateOfBirth', date_of_birth))
-    if state:
-      code_answers.append(('state', Concept(PPI_SYSTEM, state)))
-    qr = make_questionnaire_response_json(
-        participant_id,
-        questionnaire_id,
-        code_answers=code_answers,
-        date_answers=date_answers)
-    self.send_post('Participant/%s/QuestionnaireResponse' % participant_id, qr)
-
-  def _submit_consent_questionnaire_response(
-      self, participant_id, questionnaire_id, ehr_consent_answer):
-    code_answers = [('ehrConsent', Concept(PPI_SYSTEM, ehr_consent_answer))]
-    qr = make_questionnaire_response_json(
-        participant_id, questionnaire_id, code_answers=code_answers)
-    self.send_post('Participant/%s/QuestionnaireResponse' % participant_id, qr)
-
-  def _submit_empty_questionnaire_response(self, participant_id,
-                                           questionnaire_id):
-    qr = make_questionnaire_response_json(participant_id, questionnaire_id)
-    self.send_post('Participant/%s/QuestionnaireResponse' % participant_id, qr)
 
   def _create_data(self):
     HPODao().insert(HPO(hpoId=PITT_HPO_ID + 1, name='AZ_TUCSON'))
@@ -92,22 +57,22 @@ class PublicMetricsExportTest(CloudStorageSqlTestBase, FlaskTestBase):
     questionnaire_id_2 = self.create_questionnaire('questionnaire4.json')
     questionnaire_id_3 = self.create_questionnaire(
         'all_consents_questionnaire.json')
+
     with FakeClock(TIME):
-      participant = Participant(
+      participant = self._participant_with_defaults(
           participantId=1,
+          version=2,
           biobankId=2,
-          providerLink=make_primary_provider_link_for_name('AZ_TUCSON'))
+          providerLink=make_primary_provider_link_for_name('PITT'))
       participant_dao.insert(participant)
       self.send_consent('P1', email='bob@gmail.com')
 
-    with FakeClock(TIME):
       # Participant 2 starts out unpaired; later gets paired automatically when their physical
       # measurements come in.
       participant2 = Participant(participantId=2, biobankId=3)
       participant_dao.insert(participant2)
       self.send_consent('P2', email='bob@fexample.com')
 
-    with FakeClock(TIME):
       # Test HPO affiliation; this test participant is ignored.
       participant3 = Participant(
           participantId=3,
@@ -131,27 +96,12 @@ class PublicMetricsExportTest(CloudStorageSqlTestBase, FlaskTestBase):
       participant_dao.insert(participant5)
       self.send_consent('P5', email='ch@gmail.com')
 
-    with FakeClock(TIME_2):
-      # This update to participant has no effect, as the HPO ID didn't change.
-      participant = self._participant_with_defaults(
-          participantId=1,
-          version=1,
-          biobankId=2,
-          providerLink=make_primary_provider_link_for_name('AZ_TUCSON'))
-      participant_dao.update(participant)
       self.submit_questionnaire_response('P1', questionnaire_id,
                                          RACE_WHITE_CODE, 'male', None,
                                          datetime.date(1980, 1, 2))
       self.submit_questionnaire_response(
           'P2', questionnaire_id, RACE_NONE_OF_THESE_CODE, None, None, None)
 
-    with FakeClock(TIME_3):
-      participant = self._participant_with_defaults(
-          participantId=1,
-          version=2,
-          biobankId=2,
-          providerLink=make_primary_provider_link_for_name('PITT'))
-      participant_dao.update(participant)
       self.send_post('Participant/P2/PhysicalMeasurements',
                      load_measurement_json(2))
       self.send_post('Participant/P2/BiobankOrder', load_biobank_order_json(2))
@@ -163,9 +113,9 @@ class PublicMetricsExportTest(CloudStorageSqlTestBase, FlaskTestBase):
                                          None)
       self.submit_questionnaire_response('P2', questionnaire_id_2, None, None,
                                          'PIIState_VA', None)
-      self._submit_consent_questionnaire_response('P1', questionnaire_id_3,
+      self.submit_consent_questionnaire_response('P1', questionnaire_id_3,
                                                   CONSENT_PERMISSION_NO_CODE)
-      self._submit_consent_questionnaire_response('P2', questionnaire_id_3,
+      self.submit_consent_questionnaire_response('P2', questionnaire_id_3,
                                                   CONSENT_PERMISSION_YES_CODE)
       sample_dao = BiobankStoredSampleDao()
       sample_dao.insert(
@@ -173,13 +123,13 @@ class PublicMetricsExportTest(CloudStorageSqlTestBase, FlaskTestBase):
               biobankStoredSampleId='abc',
               biobankId=2,
               test='test',
-              confirmed=TIME_2))
+              confirmed=TIME))
       sample_dao.insert(
           BiobankStoredSample(
               biobankStoredSampleId='def',
               biobankId=3,
               test='1SAL',
-              confirmed=TIME_2))
+              confirmed=TIME))
 
   def test_metric_export(self):
     self._create_data()
